requirements.txt
Django==5.2.1
dj-rest-auth==7.0.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
requests==2.32.3
gunicorn==23.0.0
django-allauth==65.8.0
django-environ==0.12.0
python-decouple==3.8
psycopg2-binary==2.9.10
whitenoise==6.9.0